// src/services/types/ingredient.ts
export type TIngredient = {
	_id: string;
	name: string;
	type: string;
	proteins: number;
	fat: number;
	carbohydrates: number;
	calories: number;
	price: number;
	image: string;
	image_mobile: string;
	image_large: string;
	uid?: string;
};

export type TIngredientState = {
	details: TIngredient | null;
};